Renaming SelectedObjects to Selection
