+++ lib/SelectionProvider/index.js
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import PropTypes from 'prop-types';
 import ShortcutTrigger from '../ShortcutTrigger';
 
-const SelectedObjectsContext = React.createContext({});
+const SelectionContext = React.createContext({});
 
-const SelectedObjectsProvider = ({ children }) => {
+const SelectionProvider = ({ children }) => {
   const [selected, setSelected] = useState([]);
   const [intendedToSelect, setIntendedToSelect] = useState([]);
   const [currentSelectedType, setCurrentSelectedType] = useState(null);
@@ -58,15 +58,15 @@ const SelectedObjectsProvider = ({ children }) => {
     intendedToSelect
   };
   return (
-    <SelectedObjectsContext.Provider value={sharedState}>
+    <SelectionContext.Provider value={sharedState}>
       {children}
       <ShortcutTrigger shortcutKey="Escape" onShortcutTrigger={clear} />
-    </SelectedObjectsContext.Provider>
+    </SelectionContext.Provider>
   );
 };
 
-SelectedObjectsProvider.propTypes = {
+SelectionProvider.propTypes = {
   children: PropTypes.node.isRequired
 };
 
-export { SelectedObjectsProvider, SelectedObjectsContext };
+export { SelectionProvider, SelectionContext };
