+++ lib/SelectionProvider/stories/SelectionStory.js
@@ -1,6 +1,6 @@
 import React, { Fragment } from 'react';
 import { storiesOf } from '@storybook/react';
-import { SelectedObjectsProvider, SelectedObjectsContext } from '..';
+import { SelectionProvider, SelectionContext } from '..';
 import Button from '../../Button';
 import StoryItem from '../../../stories/styleguide/StoryItem';
 
@@ -14,8 +14,8 @@ const items = [...new Array(8)].map((i, index) => ({
 stories.add('Selected Objects Provider', () => (
   <>
     <StoryItem title="Selecting Items">
-      <SelectedObjectsProvider>
-        <SelectedObjectsContext.Consumer>
+      <SelectionProvider>
+        <SelectionContext.Consumer>
           {context => (
             <Fragment>
               {items.map(i => (
@@ -30,13 +30,13 @@ stories.add('Selected Objects Provider', () => (
               </Button>
             </Fragment>
           )}
-        </SelectedObjectsContext.Consumer>
-      </SelectedObjectsProvider>
+        </SelectionContext.Consumer>
+      </SelectionProvider>
     </StoryItem>
 
     <StoryItem title="Restricting items using currentSelectedType">
-      <SelectedObjectsProvider>
-        <SelectedObjectsContext.Consumer>
+      <SelectionProvider>
+        <SelectionContext.Consumer>
           {context => (
             <Fragment>
               {items.map(i => (
@@ -58,8 +58,8 @@ stories.add('Selected Objects Provider', () => (
               </Button>
             </Fragment>
           )}
-        </SelectedObjectsContext.Consumer>
-      </SelectedObjectsProvider>
+        </SelectionContext.Consumer>
+      </SelectionProvider>
     </StoryItem>
   </>
 ));
