+++ lib/SelectionProvider/useObjectSelector.js
@@ -1,11 +1,11 @@
 import { useContext } from 'react';
-import { SelectedObjectsContext } from './index';
+import { SelectionContext } from './index';
 
 export function useObjectSelector(
   id,
   type,
   idsToSelect,
-  disabledCallback = () => false
+  isDisabledChecker = () => false
 ) {
   const {
     selected,
@@ -15,10 +15,10 @@ export function useObjectSelector(
     currentSelectedType,
     intendedToSelect,
     setIntendedToSelect
-  } = useContext(SelectedObjectsContext);
+  } = useContext(SelectionContext);
 
   const isSelected = selected.indexOf(id) !== -1;
-  const isDisabled = disabledCallback(currentSelectedType, isSelected);
+  const isDisabled = isDisabledChecker(currentSelectedType, isSelected);
   const isHovered = intendedToSelect.indexOf(id) !== -1;
 
   const getIds = (additionalIds = []) => [
