+++ lib/index.js
@@ -97,9 +97,6 @@ export ImageLoader from './images/ImageLoader';
 export InputWithButton from './InputWithButton';
 export Pill from './Pill';
 export GuideCard from './GuideCard';
-export {
-  SelectedObjectsProvider,
-  SelectedObjectsContext
-} from './SelectedObjectsProvider';
+export { SelectionProvider, SelectionContext } from './SelectionProvider';
 export { DndProvider, DragPreview, Draggable, Droppable } from './DnD';
 export { InViewList, InViewListItem, WindowingProvider } from './Windowing';
