+++ stories/index.js
@@ -61,6 +61,6 @@ import Pill from './components/Pill';
 import GuideCard from '../lib/GuideCard/stories/GuideCardStory';
 import DnD from '../lib/DnD/stories/DnDStory';
 import Windowing from '../lib/Windowing/stories/WindowingStory';
-import SelectedObjects from '../lib/SelectedObjectsProvider/stories/SelectedObjectsStory'
+import Selection from '../lib/SelectionProvider/stories/SelectionStory';
 
 import Hierarchy from './webapp/hierarchy/HierarchyStory';
